Move node logins listing to roles file
