+++ lib/services/role_test.go
@@ -2841,6 +2841,265 @@ func TestRoleSetEnumerateDatabaseUsers(t *testing.T) {
 	}
 }
 
+func TestEnumerateTestLogins(t *testing.T) {
+	devEnvRole := &types.RoleV5{
+		Spec: types.RoleSpecV5{
+			Allow: types.RoleConditions{
+				Logins:     []string{"devuser"},
+				Namespaces: []string{apidefaults.Namespace},
+				NodeLabels: types.Labels{
+					"env": []string{"dev"},
+				},
+			},
+		},
+	}
+
+	prodEnvRole := &types.RoleV5{
+		Spec: types.RoleSpecV5{
+			Allow: types.RoleConditions{
+				Logins:     []string{"produser"},
+				Namespaces: []string{apidefaults.Namespace},
+				NodeLabels: types.Labels{
+					"env": []string{"prod"},
+				},
+			},
+		},
+	}
+
+	anyEnvRole := &types.RoleV5{
+		Spec: types.RoleSpecV5{
+			Allow: types.RoleConditions{
+				Logins:     []string{"anyenvrole"},
+				Namespaces: []string{apidefaults.Namespace},
+				NodeLabels: types.Labels{
+					"env": []string{"*"},
+				},
+			},
+		},
+	}
+
+	rootUser := &types.RoleV5{
+		Spec: types.RoleSpecV5{
+			Allow: types.RoleConditions{
+				Logins:     []string{"root"},
+				Namespaces: []string{apidefaults.Namespace},
+				NodeLabels: types.Labels{
+					"*": []string{"*"},
+				},
+			},
+		},
+	}
+
+	roleWithMultipleLabels := &types.RoleV5{
+		Spec: types.RoleSpecV5{
+			Allow: types.RoleConditions{
+				Logins:     []string{"multiplelabelsuser"},
+				Namespaces: []string{apidefaults.Namespace},
+				NodeLabels: types.Labels{
+					"region": []string{"*"},
+					"env":    []string{"dev"},
+				},
+			},
+		},
+	}
+
+	tt := []struct {
+		name           string
+		server         types.Server
+		roleSet        RoleSet
+		expectedLogins []string
+	}{
+		{
+			name: "env dev login is added",
+			server: makeTestServer(map[string]string{
+				"env": "dev",
+			}),
+			roleSet:        NewRoleSet(devEnvRole),
+			expectedLogins: []string{"devuser"},
+		},
+		{
+			name: "env prod login is added",
+			server: makeTestServer(map[string]string{
+				"env": "prod",
+			}),
+			roleSet:        NewRoleSet(prodEnvRole),
+			expectedLogins: []string{"produser"},
+		},
+		{
+			name: "only the correct login is added",
+			server: makeTestServer(map[string]string{
+				"env": "prod",
+			}),
+			roleSet:        NewRoleSet(prodEnvRole, devEnvRole),
+			expectedLogins: []string{"produser"},
+		},
+		{
+			name: "logins from role not authorizeds are not added",
+			server: makeTestServer(map[string]string{
+				"env": "staging",
+			}),
+			roleSet:        NewRoleSet(devEnvRole, prodEnvRole),
+			expectedLogins: nil,
+		},
+		{
+			name: "role with wildcard get its logins",
+			server: makeTestServer(map[string]string{
+				"env": "prod",
+			}),
+			roleSet:        NewRoleSet(anyEnvRole),
+			expectedLogins: []string{"anyenvrole"},
+		},
+		{
+			name: "can return multiple logins",
+			server: makeTestServer(map[string]string{
+				"env": "prod",
+			}),
+			roleSet:        NewRoleSet(anyEnvRole, prodEnvRole),
+			expectedLogins: []string{"anyenvrole", "produser"},
+		},
+		{
+			name: "can return multiple logins from same role",
+			server: makeTestServer(map[string]string{
+				"env": "prod",
+			}),
+			roleSet: NewRoleSet(&types.RoleV5{
+				Spec: types.RoleSpecV5{
+					Allow: types.RoleConditions{
+						Logins:     []string{"role1", "role2", "role3"},
+						Namespaces: []string{apidefaults.Namespace},
+						NodeLabels: types.Labels{
+							"env": []string{"*"},
+						},
+					},
+				},
+			}),
+			expectedLogins: []string{"role1", "role2", "role3"},
+		},
+		{
+			name: "works with user with full access",
+			server: makeTestServer(map[string]string{
+				"env": "prod",
+			}),
+			roleSet:        NewRoleSet(rootUser),
+			expectedLogins: []string{"root"},
+		},
+		{
+			name: "works with server with multiple labels",
+			server: makeTestServer(map[string]string{
+				"env":    "prod",
+				"region": "us-east-1",
+			}),
+			roleSet:        NewRoleSet(prodEnvRole),
+			expectedLogins: []string{"produser"},
+		},
+		{
+			name: "don't add login from unrelated labels",
+			server: makeTestServer(map[string]string{
+				"env": "dev",
+			}),
+			roleSet: NewRoleSet(&types.RoleV5{
+				Spec: types.RoleSpecV5{
+					Allow: types.RoleConditions{
+						Logins:     []string{"anyregionuser"},
+						Namespaces: []string{apidefaults.Namespace},
+						NodeLabels: types.Labels{
+							"region": []string{"*"},
+						},
+					},
+				},
+			}),
+			expectedLogins: nil,
+		},
+		{
+			name: "works with roles with multiple labels that role shouldn't access",
+			server: makeTestServer(map[string]string{
+				"env": "dev",
+			}),
+			roleSet:        NewRoleSet(roleWithMultipleLabels),
+			expectedLogins: nil,
+		},
+		{
+			name: "works with roles with multiple labels that role shouldn access",
+			server: makeTestServer(map[string]string{
+				"env":    "dev",
+				"region": "us-west-1",
+			}),
+			roleSet:        NewRoleSet(roleWithMultipleLabels),
+			expectedLogins: []string{"multiplelabelsuser"},
+		},
+		{
+			name: "works with roles with regular expressions",
+			server: makeTestServer(map[string]string{
+				"region": "us-west-1",
+			}),
+			roleSet: NewRoleSet(&types.RoleV5{
+				Spec: types.RoleSpecV5{
+					Allow: types.RoleConditions{
+						Logins:     []string{"rolewithregexpuser"},
+						Namespaces: []string{apidefaults.Namespace},
+						NodeLabels: types.Labels{
+							"region": []string{"^us-west-1|eu-central-1$"},
+						},
+					},
+				},
+			}),
+			expectedLogins: []string{"rolewithregexpuser"},
+		},
+		{
+			name: "works with denied roles",
+			server: makeTestServer(map[string]string{
+				"env": "dev",
+			}),
+			roleSet: NewRoleSet(devEnvRole, &types.RoleV5{
+				Spec: types.RoleSpecV5{
+					Deny: types.RoleConditions{
+						Logins:     []string{"devuser"},
+						Namespaces: []string{apidefaults.Namespace},
+						NodeLabels: types.Labels{
+							"env": []string{"*"},
+						},
+					},
+				},
+			}),
+			expectedLogins: nil,
+		},
+		{
+			name: "works with denied roles of unrelated labels",
+			server: makeTestServer(map[string]string{
+				"env": "dev",
+			}),
+			roleSet: NewRoleSet(devEnvRole, &types.RoleV5{
+				Spec: types.RoleSpecV5{
+					Deny: types.RoleConditions{
+						Logins:     []string{"devuser"},
+						Namespaces: []string{apidefaults.Namespace},
+						NodeLabels: types.Labels{
+							"region": []string{"*"},
+						},
+					},
+				},
+			}),
+			expectedLogins: nil,
+		},
+	}
+	for _, tc := range tt {
+		t.Run(tc.name, func(t *testing.T) {
+			logins := tc.roleSet.EnumerateServerLogins(tc.server)
+			require.Equal(t, tc.expectedLogins, logins.Allowed())
+		})
+	}
+}
+
+// makeTestServer creates a server with labels and an empty spec.
+// It panics in case of an error. Used only for testing
+func makeTestServer(labels map[string]string) types.Server {
+	s, err := types.NewServerWithLabels("server", types.KindNode, types.ServerSpecV2{}, labels)
+	if err != nil {
+		panic(err)
+	}
+	return s
+}
+
 func TestCheckDatabaseNamesAndUsers(t *testing.T) {
 	roleEmpty := &types.RoleV5{
 		Metadata: types.Metadata{Name: "roleA", Namespace: apidefaults.Namespace},
